fellow_kids/requirements.txt
aiodns==2.0.0
aiohttp==3.6.2
aiosqlite3==0.3.0
async-timeout==3.0.1
attrs==19.3.0
cchardet==2.1.6
cffi==1.14.0
chardet==3.0.4
discord.py==1.3.3
idna==2.10
lavalink==3.1.2
multidict==4.7.6
Pillow==7.2.0
pycares==3.1.1
pycparser==2.20
python-dotenv==0.14.0
websockets==8.1
yarl==1.4.2
-e git+https://github.com/Rapptz/discord-ext-menus#egg=discord.ext.menus